streamlit/requirements.txt
# Streamlit frontend
streamlit>=1.30.0
streamlit_autorefresh

# Data processing
pandas>=2.2.2
plotly>=5.10.0
requests>=2.28.0
python-dotenv

# Flask + SocketIO backend
flask>=2.3.2
flask-socketio>=5.3.2
python-socketio>=5.7.1
eventlet>=0.33.3

# Optional: client-side SocketIO (if you run a Python client)
python-socketio[client]>=5.7.1
